создание контроллера и перенос в него всей логики
